Livestreaming (#909)
* scheduled recordings playlists init + selections

* make livestreaming scheduling possible + bulk actions and configs

* find, process, display, manage  and redirect livestreams

* live-events widget implementation

* fix: empty action column of video row in dozent view

* fix: add video to course playlist sql query optimization.

* prevent remove video from a playlist

* add info messages on remove livestreams from playlists (single&bulk)

* ordering migrations after master merge

* fix: livestream scheduling explicit checker

* add: livestream scheduling checker

* add light preview image

* cover playlist mapping for scheduled recordings

---------

+++ lib/Routes/Playlist/PlaylistRemoveVideo.php
@@ -32,6 +32,16 @@ class PlaylistRemoveVideo extends OpencastController
             throw new \AccessDeniedException();
         }
 
+        // Prevent removing video from playlist when it is livestream.
+        if ((bool) $video->is_livestream) {
+            return $this->createResponse([
+                'message' => [
+                    'type' => 'error',
+                    'text' => _('Entfernung des Livestream-Videos aus der Wiedergabeliste ist nicht erlaubt.')
+                ],
+            ], $response->withStatus(403));
+        }
+
         PlaylistVideos::deleteBySQL('playlist_id = ? AND video_id = ?', [
             $playlist->id, $video->id
         ]);
